First commit consisting of paginated_api for javascript
This is the baseline code for paginated APIs.  It is how we
might walk an API in the past.  It includes code for a server
written in go.  The intent of this code is to show that
separating concerns in older javascript was difficult.
This will pair with a blog article on www.tahoemph.blog which
discussed this code and will be followed by better versions.